+++ test/decorateSrcResHeaders.test.ts
@@ -9,7 +9,7 @@ describe("when userResHeaderDecorator is defined", () => {
       res.json(req.headers);
     });
     const targetServer = target.listen();
-    const targetPort = (targetServer.listener.addr as Deno.NetAddr).port;
+    const targetPort = (targetServer.addrs[0] as Deno.NetAddr).port;
 
     const app = opine();
     app.use(proxy(`http://localhost:${targetPort}`, {