+++ test/deps.ts
@@ -1,3 +1,3 @@
-export * from "https://deno.land/x/opine@1.9.1/mod.ts";
+export * from "https://deno.land/x/opine@2.0.2/mod.ts";
 export { expect } from "https://deno.land/x/expect@v0.2.9/mod.ts";
-export { superdeno } from "https://deno.land/x/superdeno@4.7.1/mod.ts";
+export { superdeno } from "https://deno.land/x/superdeno@4.7.2/mod.ts";