+++ test/status.test.ts
@@ -12,7 +12,7 @@ describe("proxies status code", () => {
       });
 
       const targetServer = target.listen();
-      const targetPort = (targetServer.listener.addr as Deno.NetAddr).port;
+      const targetPort = (targetServer.addrs[0] as Deno.NetAddr).port;
 
       const proxyServer = opine();
       proxyServer.use(proxy(`http://localhost:${targetPort}`));