+++ test/url.test.ts
@@ -26,7 +26,7 @@ describe("url: string", () => {
       };
 
       const target = proxyTarget({ handlers: [proxyRouteFn] });
-      const targetPort = (target.listener.addr as Deno.NetAddr).port;
+      const targetPort = (target.addrs[0] as Deno.NetAddr).port;
 
       const app = opine();
       app.use("/somePath/", proxy(`http://localhost:${targetPort}`));
@@ -58,7 +58,7 @@ describe("url: URL", () => {
       };
 
       const target = proxyTarget({ handlers: [proxyRouteFn] });
-      const targetPort = (target.listener.addr as Deno.NetAddr).port;
+      const targetPort = (target.addrs[0] as Deno.NetAddr).port;
 
       const app = opine();
       app.use("/somePath/", proxy(new URL(`http://localhost:${targetPort}`)));