+++ .github/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## [3.0.1] - 10-01-2022
+
+- [#12] `method` should have type `string` not `"string"`
+- feat: update to Deno `1.17.2`, std `0.120.0`
+- feat: upgrade Opine to `2.0.2` in tests
+- feat: other minor dependency upgrades
+- chore: changes to reflect GitHub repo transfer
+
 ## [3.0.0] - 21-11-2021
 
 - [#10] Store and pass the proxy response data as an `Uint8Array` instead of
@@ -7,8 +15,8 @@
 
 This impacts:
 
-- [`srcResDecorator`](https://github.com/asos-craigmorten/opine-http-proxy/tree/main#srcresdecoratorreq-res-proxyres-proxyresdata-supports-promise)
-- [`filterRes`](https://github.com/asos-craigmorten/opine-http-proxy/tree/main#filterresproxyres-proxyresdata-supports-promise-form)
+- [`srcResDecorator`](https://github.com/cmorten/opine-http-proxy/tree/main#srcresdecoratorreq-res-proxyres-proxyresdata-supports-promise)
+- [`filterRes`](https://github.com/cmorten/opine-http-proxy/tree/main#filterresproxyres-proxyresdata-supports-promise-form)
 
 Where the `proxyResData` argument will be of type `Uint8Array|null` and not
 `string|null`. If you require the value to be a string, you will need to decode